+++ controller/src/main/java/com/linbit/linstor/core/apicallhandler/controller/autoplacer/AutoUnplacer.java
@@ -349,7 +349,10 @@ public class AutoUnplacer
                 highestScore = value;
                 rscsWithHighestScore.clear();
             }
-            rscsWithHighestScore.add(rsc);
+            if (highestScore == value)
+            {
+                rscsWithHighestScore.add(rsc);
+            }
         }
 
         return rscsWithHighestScore;